+++ babel.config.js
@@ -1,17 +1,8 @@
 const path = require('path');
 
-module.exports = {
-    presets: [
-        [
-            '@babel/env',
-            {
-                modules: process.env.BABEL_ENV === 'es' ? false : 'auto',
-            },
-        ],
-    ],
+const config = {
+    presets: [],
     plugins: [
-        '@babel/plugin-proposal-object-rest-spread',
-        '@babel/plugin-transform-runtime',
         [
             'babel-plugin-custom-import-path-transform',
             {
@@ -19,5 +10,25 @@ module.exports = {
             },
         ],
     ],
-    ignore: ['**/__tests__/'],
+    ignore: ['**/__tests__/', '**/*.test.js'],
 };
+
+if (process.env.BABEL_ENV === 'es') {
+    config.presets.push([
+        '@babel/modules',
+        {
+            loose: true,
+        },
+    ]);
+} else {
+    config.presets.push([
+        '@babel/env',
+        {
+            loose: true,
+        },
+    ]);
+
+    config.plugins.push(...['@babel/plugin-proposal-object-rest-spread', '@babel/plugin-transform-runtime']);
+}
+
+module.exports = config;