+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.0.0-beta.7 - 2019-11-15
+
+### Update
+
+-   Upgrade dependencies.
+-   Update babel config:
+    -   Use `@babel/preset-modules` -> decreased package size.
+    -   Enable `loose` option -> decreased package size.
+
 ## 3.0.0-beta.6 - 2019-10-22
 
 ### Updated